Split `HtmlRuby` into its own file and export it
