+++ packages/core/lib/src/internal/core_ops.dart
@@ -1,5 +1,4 @@
 import 'dart:convert';
-import 'dart:math';
 import 'dart:typed_data';
 
 import 'package:flutter/gestures.dart';
